Fixed Bugs.
