+++ src/zTests/Misc/0HelloWorld.txt
@@ -1,0 +1,5 @@
+class Main inherits IO {
+   main(): SELF_TYPE {
+	out_string("Hello, World.\n")
+   };
+};
