+++ src/zTests/Misc/0Simple.txt
@@ -1,0 +1,13 @@
+class Main{
+    a: A;
+
+    main(): A {
+        a <- new A
+    };
+};
+
+class A {
+    method(): Int{
+        1
+    };
+};
