+++ src/zTests/Misc/10.txt
@@ -1,0 +1,10 @@
+
+class Main{
+    a: A;		
+    main(b:Int): A {
+	b <- 3
+        a
+    };
+};
+
+class A{};
