+++ src/zTests/Misc/11.txt
@@ -1,0 +1,10 @@
+class Silly 
+{ 
+	capy() : SELF_TYPE { self }; 
+};
+class Sally inherits Silly { };
+
+class Main { 
+	x : Sally <- (new Sally).capy();
+	main() : Sally { x };
+};
