+++ src/zTests/Misc/8.txt
@@ -1,0 +1,34 @@
+class H
+{
+	a:int;
+};
+
+class Z inherits H
+{
+	x:int;
+};
+
+class A inherits F
+{
+	b:int;
+};
+
+class B inherits A
+{
+	c:int;
+};
+
+class C inherits B
+{
+	e:int;
+};
+
+class D inherits C
+{
+	f:int;
+};
+
+class F inherits D
+{
+	g:int;
+};
